Moved redux tests to its own file
